fix: set release mode as draft in semantic-release configuration to enable attaching release assets later in the workflow (#29)
This update modifies the `.releaserc.json` file to set the
`draftRelease` option to true, allowing releases to be created as
drafts. Additionally, the milestone release workflow is enhanced with a
new step to publish the release, transitioning it from draft to
published status using the GitHub CLI. This change improves the release
process by providing more control over the visibility of new releases
before they are officially published.

Close #28